ARTEMIS-2682: OpenWireMessageConverter refactoring
